using hilt with view model
