+++ app/src/main/java/com/plcoding/ktorclientandroid/App.kt
@@ -1,0 +1,7 @@
+package com.plcoding.ktorclientandroid
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App: Application()
