--- app/src/main/java/com/plcoding/ktorclientandroid/MainActivity.kt
@@ -1,54 +1,0 @@
-package com.plcoding.ktorclientandroid
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.produceState
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.plcoding.ktorclientandroid.data.remote.PostsService
-import com.plcoding.ktorclientandroid.data.remote.dto.PostResponse
-import com.plcoding.ktorclientandroid.ui.theme.KtorClientAndroidTheme
-
-class MainActivity : ComponentActivity() {
-
-    private val service = PostsService.create()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            val posts = produceState<List<PostResponse>>(
-                initialValue = emptyList(),
-                producer = {
-                    value = service.getPosts()
-                }
-            )
-            KtorClientAndroidTheme {
-                Surface(color = MaterialTheme.colors.background) {
-                    LazyColumn {
-                        items(posts.value) {
-                            Column(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(16.dp)
-                            ) {
-                                Text(text = it.title, fontSize = 20.sp)
-                                Spacer(modifier = Modifier.height(4.dp))
-                                Text(text = it.body, fontSize = 14.sp)
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
