+++ app/src/main/java/com/plcoding/ktorclientandroid/di/ApiModule.kt
@@ -1,0 +1,33 @@
+package com.plcoding.ktorclientandroid.di
+
+import com.plcoding.ktorclientandroid.data.remote.PostsService
+import com.plcoding.ktorclientandroid.data.remote.PostsServiceImplementation
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import io.ktor.client.*
+import io.ktor.client.engine.android.*
+import io.ktor.client.features.json.*
+import io.ktor.client.features.json.serializer.*
+import io.ktor.client.features.logging.*
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ApiModule {
+    @Provides
+    @Singleton
+    fun providePostsService(client: HttpClient): PostsService = PostsServiceImplementation(client)
+
+    @Provides
+    @Singleton
+    fun providesHttpClient(): HttpClient = HttpClient(Android) {
+        install(Logging) {
+            level = LogLevel.ALL
+        }
+        install(JsonFeature) {
+            serializer = KotlinxSerializer()
+        }
+    }
+}
