+++ app/src/main/java/com/plcoding/ktorclientandroid/ui/MainActivity.kt
@@ -1,0 +1,24 @@
+package com.plcoding.ktorclientandroid.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import com.plcoding.ktorclientandroid.ui.screens.HomeScreen
+import com.plcoding.ktorclientandroid.ui.theme.KtorClientAndroidTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            KtorClientAndroidTheme {
+                Surface(color = MaterialTheme.colors.background) {
+                    HomeScreen()
+                }
+            }
+        }
+    }
+}
