+++ build.gradle
@@ -2,6 +2,7 @@
 buildscript {
     ext {
         compose_version = '1.0.1'
+        hilt_version = '2.38.1'
     }
     repositories {
         google()
@@ -11,6 +12,8 @@ buildscript {
         classpath "com.android.tools.build:gradle:7.0.2"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21"
         classpath "org.jetbrains.kotlin:kotlin-serialization:1.5.21"
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
